pyproject.toml
[tool.poetry]
name = "hubspotldapsync"
version = "0.1.0"
description = "Sync Hubspot contacts with LDAP servers"
authors = ["Matthew Macdonald-Wallace <matt@doics.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hubspot-api-client = "^7.5.0"
python-dotenv = "^1.0.0"
python-ldap = "^3.4.3"
ldap3 = "^2.9.1"
gql = "^3.4.1"
aiohttp = "^3.8.4"
requests = "^2.30.0"
requests-toolbelt = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
